requirements.txt
black
flake8
isort
mypy
pytest
uvicorn
fastapi
pydantic
MyApplication
python-dotenv
typer
Faker
starlette
httpx
types-requests
requests